BJS2 - 17597 Интеграция с генератором случайных аватаров юзера
+ Разрешил конфликты
