--- src/main/java/school/faang/user_service/controller/user/UserController.java
@@ -1,24 +1,0 @@
-package school.faang.user_service.controller.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import school.faang.user_service.dto.UserDto;
-import school.faang.user_service.service.UserService;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/users")
-public class UserController {
-
-    private final UserService userService;
-
-    @PostMapping("/byList")
-    List<UserDto> getUsersByIds(@RequestBody List<Long> ids) {
-        return userService.getUsersDtoByIds(ids);
-    }
-}
