+++ src/main/resources/application.yaml
@@ -43,13 +43,13 @@ payment-service:
   port: 9080
 
 aws:
-  accessKey: accessKey
-  secretKey: secretKey
-  region: region
+  accessKey: dummy-access-key
+  secretKey: dummy-secret-key
+  region: eu-north-1
 
 services:
   s3:
-    bucket-name: bucketName
+    bucket-name: dummy-bucket-name
     isMocked: true
   avatar:
     pattern: avatar_%d.${dice-bear.file-type}
@@ -59,7 +59,12 @@ services:
     seed-range: 20000
     content-type: image/jpeg
 
-springfox.documentation.swagger.v2.path: /api-docs
+springfox:
+  documentation:
+    swagger:
+      v2:
+       path: /api-docs
+
 dice-bear:
   url: https://api.dicebear.com
   version: 9.x
