+++ src/test/java/school/faang/user_service/service/mentorship/MentorshipServiceTest.java
@@ -68,7 +68,7 @@ public class MentorshipServiceTest {
     @DisplayName("testGetMentees")
     public void testGetMentees() {
         when(mentorshipRepository.findById(mentor.getId())).thenReturn(mentorOptional);
-        when(userMapper.usersToUserDTOs(mentor.getMentees())).thenReturn(menteesDto);
+        when(userMapper.toDtoList(mentor.getMentees())).thenReturn(menteesDto);
 
         List<UserDto> result = mentorshipService.getMentees(mentor.getId());
 
@@ -81,7 +81,7 @@ public class MentorshipServiceTest {
     @DisplayName("testGetMentors")
     public void testGetMentors() {
         when(mentorshipRepository.findById(mentee.getId())).thenReturn(menteeOptional);
-        when(userMapper.usersToUserDTOs(mentee.getMentors())).thenReturn(mentorsDto);
+        when(userMapper.toDtoList(mentee.getMentors())).thenReturn(mentorsDto);
 
         List<UserDto> result = mentorshipService.getMentors(mentee.getId());
